+++ server/services/deezerMusicProvider.js
@@ -87,7 +87,7 @@ export const DEEZER_GENRE_TAXONOMY = {
   electronic: { chartId: 113, searches: ['genre:"dance"', 'genre:"electro"', 'electronic dance music'], minFans: 100000, minRank: 250000 },
   latin: { chartId: 197, searches: ['genre:"latin"'], minFans: 150000, minRank: 300000 },
   kpop: { chartId: null, searches: ['k-pop', 'kpop', 'genre:"k-pop"'], minFans: 25000, minRank: 200000 },
-  anime: { chartId: null, searches: ['anime ost', 'anime opening', 'japanese animation', 'j-rock anime'], minFans: 1000, minRank: 100000 },
+  anime: { chartId: null, searches: ['anime opening theme', 'anime ending theme', 'anime ost', 'tv anime opening'], minFans: 25000, minRank: 200000 },
   gaming: { chartId: 173, searches: ['video game soundtrack', 'video game music', 'original video game score'], minFans: 1000, minRank: 150000 },
   cinematic: { chartId: 173, searches: ['film score', 'motion picture soundtrack', 'cinematic score'], minFans: 1000, minRank: 150000 },
   poppunk: { chartId: 85, searches: ['pop punk', 'midwest emo', 'emo punk'], minFans: 50000, minRank: 250000 },